Send full host-id in CSR signing request
We want to let hosts authenticate with Conjur from anywhere in the policy
tree. For this to work, we send the full username in the CSR request, instead
of just the username suffix (machine identity) like it was before.
To maintain backwards compatibility, the CSR's common-name still contains
only the username suffix (the machine identity) and the prefix is sent in
a header named "Host-Id-Prefix". This header content will be extracted in
the server and will be appended to the suffix from the CSR's common-name.

Integration tests for this change are located in:
  - Conjur server: We deploy Conjur in GKE and send `/inject_client_cert`
                   and `/authenticate` requests in different permutations.
  - kubernetes-conjur-demo: We deploy a Conjur cluster and an `authn-client`
                            pod to see hosts can authenticate with it.