Parse username during config build
The username parsing into prefix & suffix is now done when we build the
config. It makes more sense to have it these instead of passing around
the config and parsing it in another place. Also, this way it is more
relevant to do input validations, upon extraction from the environment.