Authenticate hosts without application identity in the id
Prior to this commit, we assumed that the application identity of the
host was defined in the host id, with a 3-part suffix. Now that the
Conjur server accepts authentication of hosts that have their
application identity defined in annotations, we need to modify the
client to remove the restriction on the host id. Additionally, this
commit changed the suffix from having the last 3 parts of the id to have
 the actual id, while the prefix has the policy id.

 To maintain backwards compatibility with an old Conjur server, we need
 to support a suffix that includes the 3-part application identity (and
 as mentioned earlier, the suffix should include the host-id which in
 this case is the 3-part application identity). In such a case, we take
 the last 3 parts of the host id as the suffix so it will have the
 application identity as the id.