Rearrange `publish` script to make more sense
Once we removed the internal registry pushing then it makes no sense to
have the `if` clause inside the `for` loop, and then have the same `if`
clause again.

This commit rearranges the logic so we have the `if` clause once and
then what needs to be done in that case.