Update the authenticator design
In this commit we add some notes and light edits to the original authenticator
design document to correct obvious differences between the original design and
the current implementation.

A more thorough and indepth review would still be valuable, but this should
give an accurate high level picture of the process the authenticator client
uses to provide an access token to applications running in a container env.