Fixed string representation of Username struct
When Username changed from a simple string to a struct, none of its
references changed and all output showed the raw struct. Since Suffix
and Prefix fields are redundant, the new string formatter for this
object now prints the FullUsername field only which should fix both this
repo's output and any other repo that uses this module.