Load cert from file if it exists
The Login method sends a login request to the server and then tries to
read the cert file from the path where it should be injected to by the
server.

However, in case it is still not there then we fail and retry to login
but it's possible that the file is already there from the previous
request. In this case, we don't need to send another login request to
the server and we can read it from the path.

This will reduce the load on the server as we will have less requests to
 it.