Reset expBackoff after we wait for TokenRefreshTimeout
Prior to this commit, we hit the reset button on the expBackoff clock
before waiting the token refresh timeout. This means that the next
interval will start 6 minutes (by default) into the clock and only one
cycle will run (as we exceed the max elapsed time of 2 minutes).

Resetting the clock after the timeout waits ensures that the next cycle
will run for at most 2 minutes as intended.