Move authentication logs from backoff to auth methods
Some log messages should be part of the authentication methods so that
they are shown also when other projects consume the authn library.

For example, the secrets-provider calls the method `authn.Authenticate()`.
in the current implementation, we will not log the message
`Authenticating as user '%s'` as it's written in the backoff loop which
is not consumed in the secrets provider. Moving these messages into the
methods will let the consuming projects have them.