Refactor request methods
Both ValidatedResponse and EmptyResponse are sort of doing the same
thing but their names are focusing on the wrong thing. We would benefit
more by having 2 functions - 1 that validates the responses, and 1 that
reads the body from the response (which is already implemented).

This way we can just validate the response instead of calling
EmptyResponse and validate the response + read the body instead of
calling DataResponse.