Perform all authentication flow in Authenticate
Consumers of the authn-client library currently need to first call the
`Authenticate` method, get the output and call the method
`ParseAuthenticationResponse` with it. This adds redundant complexity
to the flow and can be reduced easily by letting the `Authenticate`
method perform the whole flow by itself. The impact can be seen in
`main.go` (and will affect other consumers such as the secrets-provider
and secretless).