Perform all authentication flow in Authenticate (#179)
* Refactor request methods

Both ValidatedResponse and EmptyResponse are sort of doing the same
thing but their names are focusing on the wrong thing. We would benefit
more by having 2 functions - 1 that validates the responses, and 1 that
reads the body from the response (which is already implemented).

This way we can just validate the response instead of calling
EmptyResponse and validate the response + read the body instead of
calling DataResponse.

* Perform all authentication flow in Authenticate

Consumers of the authn-client library currently need to first call the
`Authenticate` method, get the output and call the method
`ParseAuthenticationResponse` with it. This adds redundant complexity
to the flow and can be reduced easily by letting the `Authenticate`
method perform the whole flow by itself. The impact can be seen in
`main.go` (and will affect other consumers such as the secrets-provider
and secretless).

* Add CHANGELOG entry for `Authenticate` method refactor