Remove temporary TAG value
There is no use in setting the TAG argument to `dev` as it is set in
`./bin/build`. It is more accurate to leave it blank and say that it is
set in `./bin/build` rather than setting it to some value and saying
that the value will change.

This change will reduce understandment complexity.