Updating get-conjur-cert.sh
This fixes the following issues:

When using the -f option if the path is relative, the file destination is relative
to the script directory, not the users directory.
For the external case, it would be nice to check if openssl was installed to tell
the user why it's failing.
Depending on the kubectl version, kubectl run might make a depoyment or pod,
it should always be deployment.
Verify should work if there is a http:// prefix on the URL or not.