Add Helm authentication validation test
Adds an optional Helm chart test that can validate that a deployed Helm chart
release is actually capable of authenticating with the configured Conjur
instance. This Conjur authentication validation testing requires the
configuration of a special "validator" host ID in Conjur security policy
that supports Conjur Kubernetes authentication. This "validator" host ID
does not require access to Conjur secrets for the purposes of this
authentication test.

An example Conjur security policy that includes an "authenticate-only"
validator host ID for testing authentication can be found in
`assets/example-conjur-policy-with-validator.yaml`.

When the Conjur authentication validation test is enabled, the Helm test
Pod that is deployed for the `helm test ...` command will include an
authn-k8s client sidecar container that uses values from the Golden
ConfigMap to attempt to authenticate with the Conjur instance. Logs from the
authn-k8s client sidecar container are copied to a volume that is shared
with the Helm test's main test container, so that the test container can
scan the logs for some common error codes.