Fixes missing publish to DockerHub for new release tags
Currently, a conjur-authn-k8s-client image is not getting published
(pushed) to DockerHub when new Git tags have been added.

The issue is that the Jenkinsfile is looking for a new tag (which is
correct), but also that the branch is `main` (which is incorrect...
the branch will be the version tag, e.g. `v0.20.0`).

Also had to add a `--latest` argument in call to `bin/publish` because
this script is looking for `$1` positional parameter, and the
`bin/build_utils` script that it includes is setting `set -euo pipefail`.