Fixes broken E2E workflow scripts
After a recent change
(https://github.com/cyberark/conjur-oss-helm-chart/pull/154)
to the Kubernetes-in-Docker example in the cyberark/conjur-oss-helm-chart
repo, the GitHub actions tests that use the E2E workflow scripts started
failing in CI.

The root cause is that the example scripts in the conjur-oss-helm-chart were
reorganized to accommodate both a KinD example and an OpenShift example.
So now the E2E workflow scripts need to be modified to use the new script
paths.