Downgrade pet store app to fix broken E2E workflow tests
The E2E workflow tests on KinD are currently failing after a recent
new release of the Pet Store demo app that includes 'summon' and
'summon-conjur' in the demo Docker image. The new Pet Store image is causing a
problem when the E2E workflow attempts to build its own Docker image
by also adding 'summon' and 'summon-conjur' to this image, causing a
conflict.

This change also includes a temporary fix for failures in the Helm test
validation CI. This is failing because there are Helm test hooks applied
to the test ConfigMaps. For some reason, Helm doesn't create the Helm
test ConfigMaps for 'helm test ...' when these hooks are present, so
these hooks are being backed out for now.