Adds app that uses Secrets Provider in standalone mode to E2E scripts
This change adds the deployment of an application that uses the Secrets
Provider in standalone mode to the E2E application deployment scripts.

This change depends upon PR #373 merging, since PR #373 adds a Helm
subchart for an application that uses Secrets Provider in standalone
mode. There is one line in the `start` script that will need to be
changed after PR #373 has merged in order to enable deployment of
this application.

Addresses Issue #292