Updating the start script with platform_login for Openshift
Updating the cleanup scripts to handle the unique conjur-oss
namespace and to check if resources exist before deleting to avoid
errors when it can't find the resources.
Updated the platform_login.sh script to act as a standalone script
to do a platform login as well as being called by the start script.