Fix calls to functions that use 'eval' in E2E test scripts
There are a couple of functions in 'utils.sh' file for the E2E workflow
test scripts that use the Bash 'eval' command:

- run_conjur_cmd_as_admin()
- wait_for_it()

The calls made to these two functions pass in a command string, and
these functions in turn execute the command strings using 'eval'.

In some cases, the callers to these functions include environment variable
values in the command strings that they dynamically generate. Here is
one example of a call to 'wait_for_it':

```
  wait_for_it 300 "$cli exec $conjur_cli_pod -- \
     bash -c \"
       conjur_appliance_url=${CONJUR_APPLIANCE_URL} \
       CONJUR_ACCOUNT=${CONJUR_ACCOUNT} \
       CONJUR_ADMIN_PASSWORD=${CONJUR_ADMIN_PASSWORD} \
       DB_PASSWORD=${SAMPLE_APP_BACKEND_DB_PASSWORD} \
       TEST_APP_NAMESPACE_NAME=${TEST_APP_NAMESPACE_NAME} \
       TEST_APP_DATABASE=${TEST_APP_DATABASE} \
       /policy/load_policies.sh
     \"
   "
```

The environment variable expressions in these statements should be
encapsulated in quotes (single or double) to prevent potential code
injection that might occur when these variables are expanded and
executed as part of the `bash -c ...` command, if these environment
variables (which are retrieved from Conjur) happen to contain executable
code.

This change also adds quotes around some bash variables that are
interpreted as strings, to adhere to Bash style guides:

https://github.com/cyberark/community/blob/main/Conjur/conventions/bash-style-guide.md#style-baseline