Fixes 'tmpcharts: file exists' errors in E2E tests
This change alters the way that the Helm dependency update for the
Secrets Provider standalone application Helm subchart is done so that it
is performed separately (i.e. serially) outside of the parallel E2E
tests that follow.

This is being done because the process of downloading of online charts
that is done as part of Helm dependency update is not thread-safe. This
download process first copies the online charts to a non-uniquely named,
temporary `tmpcharts` directory, and the download process will fail if
this `tmpcharts` directory already exists (as might happen with parallel
Jenkins test stages).

Without this change, we sometimes see the following error in Jenkins E2E
test runs:

```
Error: unable to move current charts to tmp dir: link error: cannot rename charts/app-secrets-provider-standalone/charts to charts/app-secrets-provider-standalone/tmpcharts: rename charts/app-secrets-provider-standalone/charts charts/app-secrets-provider-standalone/tmpcharts: file exists
```