Uses just-built test client image for E2E CI
This fixes the E2E Jenkins CI tests so that they use a just-built
test client image (which is used for kubectl, oc, helm, openssl, etc.
clients) when running E2E test scripts, rather than using the `edge`
version of this container that had been published on the PREVIOUS
merge to the master branch.

Without this fix, there could potentially be flaws in the current,
just-built client container image, but the CI test run will not catch
the issue. If this occurs e.g. on a merge to the master branch, then
the flawed, just-built client container image will be published as
the new `edge` version, and all subsequent test runs will be broken.

For anyone using the E2E test scripts locally/manually, the version tag
for the test client image defaults to `edge`, so that our E2E scripts
are not dependent upon having to do a build of container images. This is
convenient for those who may want to use the E2E scripts to explore
Conjur authentication.