Fixes build errors related to application postgres backend deployment
The bitnami/postgresql Helm chart that we use for deploying a postgres
backend database for our sample applications in our E2E workflow scripts
has recently had a new release published. Some of the Helm chart values
that we had been using are out-of-date, and as a result, our Helm
installs are failing.

This change updates some of the Helm chart values that we use for this
deployment, and eliminates some settings that we probably don't need
(i.e. the less "knobs" we use, the more robust we are to changes to
those "knobs").