Added automated release process to project pipeline
- Downgraded go.mod to 1.16 to match the Dockerfile version

- Added workaround for git CVE-2022-24765 in Dockerfile.helm-unit-test

- Project pipeline moved to azure-linux executor allowing public IP
access for workflow tests.

- Reconfigured bin/publish to perform releases and promotions
of contianer images based on optional flag parameters.

- Added additional functions in bin/build_utils for the automated releases.

- Maintained legacy bin/build_utils functions that are currently no longer in
use, but might provide useful in the future.

- Added go get @latest for conjur-opentelemetry-tracer upstream dependency.

- Moved temp directory creation to /bin/test-workflow/start. The temp directory
must be created by the jenkins user to prevent permission issues within
the directory. This also removes the possibility of a race condition when parallel
test stages are running and allows for developers to skip tests
during the development process.

- Added stage to push images internally to registry.tld. This allows for promoting
these images at a later time.